requirements.txt
requests==2.31.0
supabase==2.0.0
python-dotenv==1.0.0
websockets==12.0
schedule==1.2.0